requirements.txt
fastapi
uvicorn
python-whois
